60. text rendering and completetion!

+++ src/text_renderer.cpp
@@ -1,7 +1,10 @@
 #include <iostream>
 
 #include <glm/gtc/matrix_transform.hpp>
-#include <ft2build.h>
+
+#include "ft2build.h"
+#include <freetype/freetype.h>
+#include <freetype/ftoutln.h>
 #include FT_FREETYPE_H
 
 #include "text_renderer.h"
@@ -77,7 +80,7 @@ void TextRenderer::Load(std::string font, unsigned int fontSize)
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x
+            static_cast<unsigned int>(face->glyph->advance.x)
         };
         Characters.insert(std::pair<char, Character>(c, character));
     }
